Mifflin Equation page is showing results

+++ CalCount/CalCount/equations.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-func mifflin(sex: Bool, weight: Float, height: Float, age: Float) -> Float{
+func mifflin(sex: Bool, weight: Double, height: Double, age: Double) -> Double{
     if (sex == true) {
         let result = 5 + 10 * weight + 6.25 * height - 5 * age
         return result
@@ -18,6 +18,7 @@ func mifflin(sex: Bool, weight: Float, height: Float, age: Float) -> Float{
     }
 }
 
+/*
 func pennStateH(sex: Bool, weight: Float, height: Float, age: Float, tMax: Float) ->Float{
     let result = 0.94 * mifflin(sex: sex, weight: weight, height: height, age:age) + tMax * 186 - 6597
     return result
@@ -27,3 +28,4 @@ func pennState2010(sex: Bool, weight: Float, height: Float, age: Float, v: Float
     let result = 0.71 * mifflin(sex: sex, weight: weight, height: height, age:age) + v * 64 + tMax * 85 - 3085
     return result
 }
+*/
